% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creel}
\alias{creel}
\title{Vessel tracks and behaviour of creel fisheries}
\format{
A data frame with 2226 rows and 8 variables:
\describe{
\item{id}{vessel identification}
\item{rowid}{record number, unique for each vessel}
\item{time}{time of position}
\item{x}{Easting}
\item{y}{Northing}
\item{lon}{Longitude}
\item{lat}{Latitude}
\item{behaviour}{Vessel activity}
}
}
\source{
Mendo, Tania; Smout, Sophie; Photopoulou, Theoni; James, Mark (2019), Data
from: Identifying fishing grounds from vessel tracks: model-based inference
for small scale fisheries (\url{https://doi.org/10.5061/dryad.k80bp46})
}
\usage{
creel
}
\description{
A dataset of movement data collected every 60 sec from 5 trips by 5 different
small scale fishing vessels using creels. (Mendo, Tania; Smout, Sophie; Photopoulou, Theoni; James, Mark (2019), Data
from: Identifying fishing grounds from vessel tracks: model-based inference
for small scale fisheries)
}
\keyword{datasets}
