% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dansk_harbours}
\alias{dansk_harbours}
\title{Some harbour polygons used in association with the dansk-dataset}
\format{
A data frame with 942 rows and 2 variables:
\describe{
\item{SI_HARB}{sequential numbers denoting unique harbour id}
\item{geometry}{harbour polygons}
}
}
\source{
WKSSFGEO
}
\usage{
dansk_harbours
}
\description{
Some harbour polygons used in association with the dansk-dataset
}
\keyword{datasets}
