% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eflalo}
\alias{eflalo}
\title{Vessel logbook dataset}
\format{
A data frame with 4539 observations (rows) and 190 variables (columns):
\describe{
  \item{.ride}{Row number}
  \item{VE_REF}{Vessel id}
  \item{VE_FLT}{Fleet}
  \item{VE_COU}{Home country}
  \item{VE_LEN}{Vessel length}
  \item{VE_KW}{Vessel power}
  \item{VE_TON}{Vessel tonnage}
  \item{FT_REF}{Fishing trip reference number}
  \item{FT_DCOU}{Departure country}
  \item{FT_DHAR}{Departure harbour}
  \item{FT_DDAT}{Dparture date}
  \item{FT_DTIME}{Departurte time}
  \item{FT_LCOU}{Landing country}
  \item{FT_LHAR}{Landing harbour}
  \item{FT_LDAT}{Arrival date}
  \item{FT_LTIME}{Arrival time}
  \item{LE_ID}{Log event id}
  \item{LE_CDAT}{Catch date}
  \item{LE_STIME}{Log event start time}
  \item{LE_ETIME}{Log event end time}
  \item{LE_SLAT}{Log event start latitude}
  \item{LE_SLON}{Log event start longitude}
  \item{LE_ELAT}{Log event end latitude}
  \item{LE_ELON}{Log event end longitude}
  \item{LE_GEAR}{Gear}
  \item{LE_MSZ}{Mesh size}
  \item{LE_RECT}{ICES rectangle}
  \item{LE_DIV}{ICES division}
  \item{LE_MET}{Fishing activity metier (metier 6)}
  \item{LE_UNIT}{effort unit}
  \item{LE_EFF}{effort value}
  \item{LE_EFF_VMS}{Has VMS data signal}
  \item{LE_KG_}{Landing weight of species}
  \item{LE_EURO_}{Landing value of species}
  }
}
\source{
https://doi.org/10.1093/icesjms/fsv099
}
\usage{
eflalo
}
\description{
A dataset consisting of logbook (landings / values) registrations.  This dataset is from the  
vmstoolswith years in the dataset changed because of leap-year issues in
the original dataset
}
\keyword{datasets}
