% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferill-package.R
\docType{package}
\name{ferill-package}
\alias{ferill}
\alias{ferill-package}
\title{ferill: Convert "move"-object into tibbles}
\description{
The goal of {ferill} is to return various "move"-objects to tibbles for further downstream processing using the tidyverse lingo. And then some.
}
\author{
\strong{Maintainer}: Einar Hjörleifsson \email{einar.hjorleifsson@gmail.com}

}
\keyword{internal}
