% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_base_raster.R
\name{mb_base_raster}
\alias{mb_base_raster}
\title{A raster template}
\usage{
mb_base_raster()
}
\value{
a {terra} raster
}
\description{
Create a skeleton covering eez with no associated cell values.
Make it divisable by 8, 16,  ..., 2048
}
