% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_xyz_dynamic.R
\name{mb_rashade_xyz_dynamic}
\alias{mb_rashade_xyz_dynamic}
\title{Dynamic rayshading}
\usage{
mb_rashade_xyz_dynamic(
  xyz,
  r0,
  ping_cutoff = 5,
  file_prefix = "tmp/tmp",
  res = c(1, 2, 3, 4, 5),
  agg = c(1, 2, 4, 8, 16)
)
}
\arguments{
\item{xyz}{xxx}

\item{r0}{xxx}

\item{ping_cutoff}{xxx}

\item{file_prefix}{xxx}

\item{res}{xxx}

\item{agg}{xxx}
}
\description{
Dynamic rayshading
}
