% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rasterize_xyz.R
\name{mb_rasterize_xyz}
\alias{mb_rasterize_xyz}
\title{Rasterize xyz data}
\usage{
mb_rasterize_xyz(xyz, r0, agg = 1, fun = "mean", no_trim = TRUE)
}
\arguments{
\item{xyz}{A tiblle}

\item{r0}{A base raster}

\item{agg}{Level of aggregation (2, 4, ...)}

\item{fun}{Normally mean or count}
}
\value{
A terra raster
}
\description{
Rasterize xyz data
}
