% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_raster.R
\name{mb_rayshade_raster}
\alias{mb_rayshade_raster}
\title{Rayshade raster}
\usage{
mb_rayshade_raster(r, zscale = 4)
}
\arguments{
\item{r}{A single banded raster ({raster} or {terra}).}
}
\value{
A RGB array
}
\description{
An opinionated rayshading of rasterized bathymetry depth data.
}
\details{
The returned object can be viewed via rayshader::plot_maps
}
