% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_raster_rgb.R
\name{mb_rayshade_raster_rgb}
\alias{mb_rayshade_raster_rgb}
\title{Rayshayd}
\usage{
mb_rayshade_raster_rgb(r)
}
\arguments{
\item{r}{A {terra} raster}
}
\value{
A rgb {terra} raster
}
\description{
Rayshayd
}
