% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_to_rgb.R
\name{mb_rayshade_to_rgb}
\alias{mb_rayshade_to_rgb}
\title{rayshade_to_rgb}
\usage{
mb_rayshade_to_rgb(shaded, original)
}
\arguments{
\item{shaded}{The input matrix, typically the output of a rayshader operation}

\item{original}{The original raster from which matrix is derived}
}
\value{
A terra-raster layer
}
\description{
Turns a matrix into a 3 band RGB raster. See also hillshader::matrix_to_raster
}
