% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rasterize_xyz.R
\name{mb_xyz_extent}
\alias{mb_xyz_extent}
\title{Get extent of xy-dataframe}
\usage{
mb_xyz_extent(x)
}
\arguments{
\item{x}{A dataframe containing x and y}
}
\value{
an "extent" vector
}
\description{
Get extent of xy-dataframe
}
