% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metier5_benthis_lookup}
\alias{metier5_benthis_lookup}
\title{Benthis metier lookup}
\format{
A data frame with 41 observations (rows) and 2 variables (columns):
\describe{
\item{metier5}{Metier as defined by ICES}
\item{benthis_metier}{Metier as defined in the benthis project}
}
}
\source{
https://doi.org/10.1093/icesjms/fsv099
}
\usage{
metier5_benthis_lookup
}
\description{
Benthis metier lookup
}
\keyword{datasets}
