% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{ms2kn}
\alias{ms2kn}
\title{ms2kn}
\usage{
ms2kn(x)
}
\arguments{
\item{x}{A numerical vector of speed in meters per second}
}
\value{
A vector, speed in knots (nautical miles per hour)
}
\description{
meters per second to knots
}
