% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{pm_trip}
\alias{pm_trip}
\title{pm_trip}
\usage{
pm_trip(m, data, radius = 10, trip = FALSE)
}
\arguments{
\item{m}{A mapdeck object, normally one just created by mapview()}

\item{data}{The new data to be added}

\item{radius}{Radius (in meters) of the point size}

\item{trip}{add trip path, default FALSE}
}
\value{

}
\description{
pm_trip
}
