% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips.R
\name{rb_add_trips}
\alias{rb_add_trips}
\title{Add trip id to VMS data}
\usage{
rb_add_trips(
  ve,
  le,
  vessel_id = vid,
  trip_id = tid,
  time = time,
  departure = T1,
  arrival = T2
)
}
\arguments{
\item{ve}{VMS dataset}

\item{le}{Logbook dataset}

\item{vessel_id}{The vessel id in both dataset (default vid)}

\item{trip_id}{The trip id variable in the logbook dataset (default tid)}

\item{time}{The time variable in the VMS dataset (default time)}

\item{departure}{Trip departure time in the logbook dataset (default T1)}

\item{arrival}{Trip arrival time in the logbook dataset (default T2)}
}
\value{
A tibble
}
\description{
Adds the trip id to the VMS data based on vms timestamps that are between logbook
departure and arrival time. Supposedly doing the same thing as vmstools::mergeEflalo2Tacsat
}
