% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips.R
\name{rb_add_trips_vt}
\alias{rb_add_trips_vt}
\title{Add trip id to VMS data}
\usage{
rb_add_trips_vt(
  ve,
  le,
  vessel_id = VE_REF,
  trip_id = FT_REF,
  time = SI_DATE,
  departure = FT_DTIME,
  arrival = FT_LTIME
)
}
\arguments{
\item{ve}{VMS dataset}

\item{le}{Logbook dataset}

\item{vessel_id}{The vessel id in both dataset (default VE_REF)}

\item{trip_id}{The trip id variable in the logbook dataset (default FT_REF)}

\item{time}{The time variable in the VMS dataset (default SI_DATE)}

\item{departure}{Trip departure time in the logbook dataset (default FT_DTIME)}

\item{arrival}{Trip arrival time in the logbook dataset (default FT_LTIME)}
}
\value{
A tibble
}
\description{
Adds the trip id to the VMS data based on vms timestamps that are between logbook
departure and arrival time. Supposedly doing the same thing as vmstools::mergeEflalo2Tacsat
}
