% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthis.R
\name{rb_benthis_width}
\alias{rb_benthis_width}
\title{Predict gear width in meters based on Benthis}
\usage{
rb_benthis_width(metier, length, kw)
}
\arguments{
\item{metier}{Metier 5}

\item{length}{Vessel length in meters}

\item{kw}{Vessel power}
}
\value{
A numeric vector in meters
}
\description{
Predict gear width in meters based on Benthis
}
