% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rb_calc_step_time}
\alias{rb_calc_step_time}
\title{Time between steps, forcing users to explicitly specify the unit}
\usage{
rb_calc_step_time(time, units = "secs")
}
\arguments{
\item{time}{time}

\item{units}{units of return value}
}
\value{
A numeric vector
}
\description{
Time between steps, forcing users to explicitly specify the unit
}
