% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traipse_augmentation.R
\name{rb_calc_track_time}
\alias{rb_calc_track_time}
\title{Track time duration}
\usage{
rb_calc_track_time(date, weight = c(0, 1), fill = TRUE)
}
\arguments{
\item{date}{date-time in POSIXct}

\item{weight}{A numerical vector of length 2 weight to apply to calculation
of mean interval rate towards and away from ping}

\item{fill}{Boolean (default TRUE)}
}
\value{
A vector of duration in seconds
}
\description{
Calculate time duration based on sequential difference of date-time input.
The unit of time duration is seconds.
}
\details{
Function is a modification of traipse::track_time allowing for calculation
of time from previous points rather than by convention the time to next
point. It also allows for puting some weighing a.la. the datacall::intvTacsat
function.
}
