% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_workflow.R
\name{rb_cap_effort}
\alias{rb_cap_effort}
\title{Cap effort}
\usage{
rb_cap_effort(lb)
}
\arguments{
\item{lb}{logbooks}
}
\value{
a tibble, supposedly with same number of rows as the input
}
\description{
Replaces missing effort with median effort and caps effort to 12 hours (gid 6)
24 hours (gid 7) and 15 hours (gid 14). If end of haul (t2) overlaps with
beginning of next haul (t1), cut the value of t2.
}
