% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_cap_iqr}
\alias{rb_cap_iqr}
\title{Cap value to the upper multiplier of the IQR range}
\usage{
rb_cap_iqr(x, multiplier = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{multiplier}{Default is 1.5}
}
\value{
A vector of same length as x
}
\description{
Replace extreme values by less extreme using interquartile range multiplier
}
