% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_cap_miller}
\alias{rb_cap_miller}
\title{Cap value to higest sequential 1.5 log step}
\usage{
rb_cap_miller(x, val = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{val}{A single value (default 1.5), log10 step values between consequitve
ordered values of x will be capped by the maximum step value.}
}
\value{
A vector of same length as x
}
\description{
Replaces extreme values by less extreme ones.
}
