% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_cap_winsorize}
\alias{rb_cap_winsorize}
\title{Winsorize}
\usage{
rb_cap_winsorize(x, probs = 0.99)
}
\arguments{
\item{x}{A numeric vector}

\item{probs}{A single value of probabilities with values in [0,1], default 0.99.}
}
\value{
A vector of same length as x
}
\description{
Replace extreme values by less extreme ones. Default is just to cap the higher
values, but if arguement p is of length 2, lower values can also be substituted.
}
