% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi_stuff.R
\name{rb_classify_mmsi}
\alias{rb_classify_mmsi}
\title{Categories MMSI}
\usage{
rb_classify_mmsi(mmsi)
}
\arguments{
\item{mmsi}{MMSI}
}
\value{
A vector with mmsi categories
}
\description{
See: https://en.wikipedia.org/wiki/Maritime_Mobile_Service_Identity
}
