% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_convert_speed}
\alias{rb_convert_speed}
\title{Convert speed between knots and metres per second}
\usage{
rb_convert_speed(x, from = c("kn", "ms"), to = c("ms", "kn"))
}
\arguments{
\item{x}{A numeric vector of speed.}

\item{from, to}{Units: \code{"kn"} (knots) or \code{"ms"} (metres per second).}
}
\value{
A numeric vector, speed in the \code{to} unit.
}
\description{
Convert speed between knots and metres per second
}
\examples{
rb_convert_speed(0:8, from = "ms", to = "kn")
}
