% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_bb.R
\name{rb_create_bbox}
\alias{rb_create_bbox}
\title{A poor-mans bounding box}
\usage{
rb_create_bbox(x)
}
\arguments{
\item{x}{A dataframe with x and y}
}
\value{
A vector
}
\description{
A poor-mans bounding box
}
