% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_d2ir_vmstools.R
\name{rb_d2ir_vmstools}
\alias{rb_d2ir_vmstools}
\title{Get ICES rectangle from coordinates}
\usage{
rb_d2ir_vmstools(lon, lat)
}
\arguments{
\item{lon}{Longitude vector}

\item{lat}{Latitude vector}
}
\value{
Returns the rectangles as a vector
}
\description{
Get the ICES rectangle from any lon-lat position, given that this position
is within the ICES region.
}
\details{
This function is the same as vmstools::icesRectangle but takes vectors
as input rather than a dataframe.

Use with caution, seems to give non-consistent results
}
