% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_define_trip}
\alias{rb_define_trip}
\title{rb_define_trip}
\usage{
rb_define_trip(d, vid = vid, hid = hid)
}
\arguments{
\item{d}{ais/vms tracks tibble}

\item{vid}{variable name containing vessel id}

\item{hid}{variable name conaining harbour id, is NA if out of harbour}
}
\value{
A tibble with one additional variable named tid, sequential positive
values indicate trip number, negative values indicate "harbour trip".
}
\description{
The input tibble needs a variable identifying vessel (default vid) and a variable that
indicates the harbour identification number (default hid) whose value indicate 
if in harbour.  If out of harbour the value is NA.
}
