% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rb_drop_in_polygons}
\alias{rb_drop_in_polygons}
\title{Points not in polygon}
\usage{
rb_drop_in_polygons(x, y)
}
\arguments{
\item{x}{object containin geometry POINT of class sf, sfc or sfg}

\item{y}{object cotaining geometry POLYGON of class sf, sfc or sfg}
}
\value{
Normally POINT object of class sf that are within object y
}
\description{
Points not in polygon
}
