% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_event}
\alias{rb_event}
\title{Event id}
\usage{
rb_event(x)
}
\arguments{
\item{x}{A vector, normally character or integer}
}
\value{
A numerical vector that labels each "discreet" event
}
\description{
Event id
}
