% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_workflow.R
\name{rb_fill_gear_width}
\alias{rb_fill_gear_width}
\title{Gear width proxy}
\usage{
rb_fill_gear_width(lb)
}
\arguments{
\item{lb}{logbooks}
}
\value{
a tibble
}
\description{
The gear width for trawl is based on the sweeps, otherwise the plow width.
If missing use the median value.
}
