% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rb_gaussian}
\alias{rb_gaussian}
\title{Fit Gaussian mixture model on vessel speed}
\usage{
rb_gaussian(d, cs = 1.96)
}
\arguments{
\item{d}{A data frame with xx}

\item{cs}{The multiplier on the sigma value of the first Gussian mode}
}
\value{
A tibble containing lower and upper speed threshold of the first Gaussian mode
}
\description{
Fit Gaussian mixture model on vessel speed
}
