% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rb_gaussian_binary_clustering}
\alias{rb_gaussian_binary_clustering}
\title{Title}
\usage{
rb_gaussian_binary_clustering(d, cs = 1.96)
}
\arguments{
\item{d}{A data frame with xx}

\item{cs}{The multiplier on the sigma value of the first Gussian mode (not active)}
}
\value{
A tibble
}
\description{
Title
}
