% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_gear_from_metier}
\alias{rb_gear_from_metier}
\title{Derive gear from metier 5 or 6}
\usage{
rb_gear_from_metier(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A vector
}
\description{
Obtain gear type (sometimes wrongly called metier 4?) from metier5 or  6
}
