% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_get_ices_gears}
\alias{rb_get_ices_gears}
\title{Get ICES gear table}
\usage{
rb_get_ices_gears(trim = TRUE, valid = TRUE)
}
\arguments{
\item{trim}{Boolean (default TRUE), return only key variables}

\item{valid}{Boolean (default TRUE), returns only metiers not depreciated.
Only actuelt if trim is TRUE.}
}
\value{
A tibble containing target list and description
}
\description{
Get ICES gear table
}
