% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_get_ices_metier5_benthis_lookup}
\alias{rb_get_ices_metier5_benthis_lookup}
\title{Get a table matching metier 5 and benthis metier}
\usage{
rb_get_ices_metier5_benthis_lookup(trim = TRUE, correct = TRUE)
}
\arguments{
\item{trim}{Boolean (default TRUE), return only key variables}

\item{correct}{Boolean (default TRUE), not active}
}
\value{
A vector
}
\description{
Extracts key variables from https://raw.githubusercontent.com/ices-eg/RCGs/master/Metiers/Reference_lists/RDB_ISSG_Metier_list.csv
}
