% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_get_ices_metier6}
\alias{rb_get_ices_metier6}
\title{Get ICES metier 6 table}
\usage{
rb_get_ices_metier6(trim = TRUE, valid = TRUE)
}
\arguments{
\item{trim}{Boolean (default TRUE), return only key variables}

\item{valid}{Boolean (default TRUE), returns only metiers not depreciated.
Only actuelt if trim is TRUE.}
}
\value{
A tibble containing metier 6 and description
}
\description{
Get ICES metier 6 table
}
