% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_identify_trip}
\alias{rb_identify_trip}
\title{rb_identify_trip}
\usage{
rb_identify_trip(d, vid = vid, hid = hid)
}
\arguments{
\item{d}{ais/vms tracks tibble}

\item{vid}{variable name containing vessel id}

\item{hid}{variable name conaining harbour id}
}
\value{
A tibble
}
\description{
rb_identify_trip
}
