% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whackies.R
\name{rb_interpolate}
\alias{rb_interpolate}
\title{A wrapper around approx}
\usage{
rb_interpolate(x, y)
}
\arguments{
\item{x}{A vector, nomally time}

\item{y}{The value to interpolate}
}
\value{
A vector with orgiginal and interpolated values, if any
}
\description{
A wrapper around approx
}
