% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_iqr}
\alias{rb_iqr}
\title{Calculate the upper IQR multiplier value}
\usage{
rb_iqr(x, multiplier = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{multiplier}{Default is 1.5}
}
\value{
The upper limit value
}
\description{
Calculate the upper IQR multiplier value
}
