% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rb_keep_in_polygons}
\alias{rb_keep_in_polygons}
\title{Points in polygon}
\usage{
rb_keep_in_polygons(x, y)
}
\arguments{
\item{x}{object containing geometry POINT of class sf, sfc or sfg}

\item{y}{object containing geometry POLYGON of class sf, sfc or sfg}
}
\value{
Normally POINT object of class sf that are within object y
}
\description{
Points in polygon
}
\note{
No serious testing in done
}
