% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_kn2ms}
\alias{rb_kn2ms}
\title{rb_kn2ms}
\usage{
rb_kn2ms(x)
}
\arguments{
\item{x}{A numerical vector of speed in knots}
}
\value{
A vector, speed in meters per second
}
\description{
knots to meters per second for those of us that forget the convertion number
}
\examples{
rb_kn2ms(seq(0, 16, by = 2))
}
\author{
Einar Hjörleifsson, \email{einar.hjorleifsson@gmail.com}
}
