% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_workflow.R
\name{rb_logbook}
\alias{rb_logbook}
\title{Read in logbooks}
\usage{
rb_logbook(con, VIDs, YEARS)
}
\arguments{
\item{con}{connection to Oracle}

\item{VIDs}{Vessel identification numbers}

\item{YEARS}{}
}
\value{
A tibble with both mobile and static tows
}
\description{
A single tibble for both mobile and static gears for particular vessel(s) and 
year(s).
}
