% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi_stuff.R
\name{rb_lookup_mmsi_country}
\alias{rb_lookup_mmsi_country}
\title{Return flag state from MMSI}
\usage{
rb_lookup_mmsi_country(mmsi, lookup)
}
\arguments{
\item{mmsi}{MMSI}

\item{lookup}{A tibble}
}
\value{
A vector
}
\description{
Return flag state from MMSI
}
