% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck.R
\name{rb_mapdeck}
\alias{rb_mapdeck}
\title{Plot trails}
\usage{
rb_mapdeck(d, col = "speed")
}
\arguments{
\item{d}{a tibble containing trails}

\item{col}{the variable that controls colour}
}
\value{
a mapdeck map
}
\description{
Plot trails
}
