% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_met5_from6}
\alias{rb_met5_from6}
\title{Derive metier 5 from metier 6}
\usage{
rb_met5_from6(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A vector
}
\description{
Basically trims all mesh etc. details from metier 6
}
