% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_miller}
\alias{rb_miller}
\title{Calculate the Miller limit}
\usage{
rb_miller(x, step_limit = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{step_limit}{A single value (default 1.5), log10 step values between consecutive
ordered values of x will be capped by the maximum step value.}
}
\value{
A value to be used as cap
}
\description{
Calculate the Miller limit
}
