% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi_stuff.R
\name{rb_mmsi_flag}
\alias{rb_mmsi_flag}
\title{Return flag state from MMSI}
\usage{
rb_mmsi_flag(mmsi, lookup)
}
\arguments{
\item{mmsi}{MMSI}

\item{lookup}{A tibble}
}
\value{
A vector
}
\description{
Return flag state from MMSI
}
