% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trails.R
\name{rb_mobile_vid}
\alias{rb_mobile_vid}
\title{Mapping between mobileid and vessel id}
\usage{
rb_mobile_vid(con)
}
\arguments{
\item{con}{oracle connection}
}
\value{
a query
}
\description{
Mapping between mobileid and vessel id
}
