% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_ms2kn}
\alias{rb_ms2kn}
\title{rb_ms2kn}
\usage{
rb_ms2kn(x)
}
\arguments{
\item{x}{A numerical vector of speed in meters per second}
}
\value{
A vector, speed in knots (nautical miles per hour)
}
\description{
meters per second to knots for those of us that forget the convertion number
}
\examples{
rb_ms2kn(0:8)
}
\author{
Einar Hjörleifsson, \email{einar.hjorleifsson@gmail.com}
}
