% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_number_runs}
\alias{rb_number_runs}
\title{Event id}
\usage{
rb_number_runs(x)
}
\arguments{
\item{x}{A vector, normally character or integer}
}
\value{
A numerical vector that labels each discreet event
}
\description{
Event id
}
