% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_pad_harbour}
\alias{rb_pad_harbour}
\title{Pad harbour id}
\usage{
rb_pad_harbour(tid, hid)
}
\arguments{
\item{tid}{trip id, possibly generated by a rb_trip}

\item{hid}{harbour id}
}
\value{
The hid vector with added harbour id to the first and the last point
of a trip.
}
\description{
Add the harbour id to the first and the last point of an actual trip
}
