% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_peek}
\alias{rb_peek}
\title{rb_peek}
\usage{
rb_peek(d, what, criteria)
}
\arguments{
\item{d}{ais/vms track tibble}

\item{what}{the variable name to check}

\item{criteria}{the critera.}
}
\value{
a tibble
}
\description{
Experimental function, returns rows adjacent to the one fulfilling a critera
}
