% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck.R
\name{rb_plot_trail}
\alias{rb_plot_trail}
\title{Plot trails}
\usage{
rb_plot_trail(
  d,
  col = "speed",
  tooltip = "speed",
  no_lines = TRUE,
  radius = 400,
  highlight_colour = "black",
  stroke_colour = "cyan"
)
}
\arguments{
\item{d}{a tibble containing trails}

\item{col}{the variable that controls colour}

\item{tooltip}{the variable to show as tooltip}

\item{no_lines}{plot track as lines (default = TRUE)}

\item{radius}{the radius of the points (default 400 m)}

\item{highlight_colour}{colour of track (default black)}

\item{stroke_colour}{stroke colour of track (default cyan)}
}
\value{
a mapdeck map
}
\description{
Plot trails
}
