% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rb_plot_trip}
\alias{rb_plot_trip}
\title{rb_plot_trip}
\usage{
rb_plot_trip(data, tid, radius = 10, col = "speed", trip = TRUE)
}
\arguments{
\item{data}{A tibble containing variable names lon and lat (crs 4326)}

\item{tid}{Trip id variable name, default tid}

\item{radius}{Radius (in meters) of the point size (default 10)}

\item{col}{What variable to be used for colour scale (default speed)}

\item{trip}{add trip path, default FALSE}
}
\value{
xxx
}
\description{
rb_plot_trip
}
