% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rb_points_in_polygons}
\alias{rb_points_in_polygons}
\title{Points in polygon}
\usage{
rb_points_in_polygons(x, y)
}
\arguments{
\item{x}{object containing geometry POINT of class sf or sfc}

\item{y}{object containing geometry POLYGON of class sf or sfc}
}
\value{
A boolean vector
}
\description{
Find points that are within polygon(s)
}
