% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_raster_rgb.R
\name{rb_rayshade_raster_rgb}
\alias{rb_rayshade_raster_rgb}
\title{Rayshayd}
\usage{
rb_rayshade_raster_rgb(r, zscale = 1, zrange = NULL)
}
\arguments{
\item{r}{A terra raster}

\item{zscale}{Value (default 1) passed to rayshader}

\item{zrange}{In NULL (default) no scaling}
}
\value{
A rgb terra raster
}
\description{
Rayshayd
}
