% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{rb_read_harbours_mfri}
\alias{rb_read_harbours_mfri}
\title{rb_read_harbours_mfri}
\usage{
rb_read_harbours_mfri()
}
\value{
An sf-tibble
}
\description{
A conenient wrapper that read a dataframe of Icelandic harbours (hid) with
associated polygon geometry from an open ftp-site.
}
