% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_workflow.R
\name{rb_read_logbook_mfri}
\alias{rb_read_logbook_mfri}
\title{Read in logbooks}
\usage{
rb_read_logbook_mfri(con, VIDs, YEARS)
}
\arguments{
\item{con}{connection to Oracle}

\item{VIDs}{Vessel identification numbers}

\item{YEARS}{xxx}
}
\value{
A tibble with both mobile and static tows
}
\description{
A single tibble for both mobile and static gears for particular vessel(s) and
year(s).
}
