% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{rb_read_survey_tracks_mfri}
\alias{rb_read_survey_tracks_mfri}
\title{rb_read_survey_tracks_mfri}
\usage{
rb_read_survey_tracks_mfri()
}
\value{
A tibble
}
\description{
A conenient wrapper that read a dataframe of Icelandic trawl vessel ais/vms
data. The variables are \code{vid} (vessel idendification numner), \code{cruise_id}
(cruise identification number) and the usual ais/vms variables (\code{time},
\code{lon}, \code{lat}, \code{speed}, \code{heading}).
}
