% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trails.R
\name{rb_read_trails}
\alias{rb_read_trails}
\title{Read all trails - redundant}
\usage{
rb_read_trails(con, VID, YEARS, use_PAM = FALSE)
}
\arguments{
\item{con}{oracle connection}

\item{VID}{vessel}

\item{YEARS}{years}
}
\value{
a tibble
}
\description{
Read all trails - redundant
}
