% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{rb_rms}
\alias{rb_rms}
\title{rb_rms}
\usage{
rb_rms(lon, lat, time)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{what}{name of variable}
}
\value{

}
\description{
The root mean square of value and the 2 upstream and 2 downstream values
}
