% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{rb_rms_tbl}
\alias{rb_rms_tbl}
\title{rb_rms}
\usage{
rb_rms_tbl(d)
}
\arguments{
\item{d}{A tibble containing lon, lat and time}
}
\value{
A tibble
}
\description{
The root mean square of value and the 2 upstream and 2 downstream values
}
\details{
rb_rms_tbl
}
