% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rb_sa}
\alias{rb_sa}
\title{Calculate step acceleration}
\usage{
rb_sa(lon, lat, time)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{time}{time in POSIXct}
}
\value{
A numerical vector
}
\description{
To use this on multiple vessels or trips, use a grouped data frame with tidyverse code like ...
}
