% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{rb_sd}
\alias{rb_sd}
\title{Distance between steps}
\usage{
rb_sd(x, y, ...)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}

\item{...}{additional arguements to geodist}
}
\value{
A distance vector in meters.
}
\description{
Calculate distance from previous point or to next point. 
Insired by traipse::track_distance but simplified and allows to pass
additional arguement to geodist::geodist.
}
