% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traipse_derivatives.R
\name{rb_speed}
\alias{rb_speed}
\title{Track speed}
\usage{
rb_speed(lon, lat, time)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{time}{datetime (POSIXct)}
}
\value{
numerical vector in nautical miles per hour
}
\description{
Calculates speed in nautical miles per hour. Based on 
traipse::track_speed but returns knots not meters per second and sets the 
first value as the second value. If the length of the vector is one, 
na values are automatically returned (rather than an error returned 
by traipse::track_speed)
}
