% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_in_not_in_polygons_filter.R,
%   R/spatial.R
\name{rb_st_keep}
\alias{rb_st_keep}
\title{Points in polygon}
\usage{
rb_st_keep(x, y)

rb_st_keep(x, y)
}
\arguments{
\item{x}{object containing geometry POINT of class sf, sfc or sfg}

\item{y}{object containing geometry POLYGON of class sf, sfc or sfg}
}
\value{
Normally POINT object of class sf that are within object y

Normally POINT object of class sf that are within object y
}
\description{
Points in polygon

Points in polygon
}
\note{
No serious tesing in done

No serious testing in done
}
