% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_in_not_in_polygons_filter.R
\name{rb_st_remove}
\alias{rb_st_remove}
\title{Points not in polygon}
\usage{
rb_st_remove(x, y)
}
\arguments{
\item{x}{object containin geometry POINT of class sf, sfc or sfg}

\item{y}{object cotaining geometry POLYGON of class sf, sfc or sfg}
}
\value{
Normally POINT object of class sf that are within object y
}
\description{
Points not in polygon
}
\note{
No serious tesing is done
}
