% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb_workflow.R
\name{rb_standardise_gear_meshsize}
\alias{rb_standardise_gear_meshsize}
\title{Standardize mesh size}
\usage{
rb_standardise_gear_meshsize(lb)
}
\arguments{
\item{lb}{logbooks}
}
\description{
Crude correction of mesh size
}
