% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_summarise_track}
\alias{rb_summarise_track}
\title{rb_summarise_track}
\usage{
rb_summarise_track(d)
}
\arguments{
\item{d}{...}
}
\value{
A summary tibble, dependent on grouping upstream
}
\description{
rb_summarise_track
}
