% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ices_vocabulary.R
\name{rb_target_from_metier}
\alias{rb_target_from_metier}
\title{Derive target from metier 5 or 6}
\usage{
rb_target_from_metier(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A vector
}
\description{
Not to be confused with metier 5
}
