% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trails.R
\name{rb_trail}
\alias{rb_trail}
\title{Read vessel trails}
\usage{
rb_trail(con, vid)
}
\arguments{
\item{con}{oracle connection}

\item{vid}{vessel id, if missing the function returns all trails}
}
\value{
a tibble
}
\description{
Read vessel trails
}
