% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{rb_trip}
\alias{rb_trip}
\title{Create trip}
\usage{
rb_trip(x)
}
\arguments{
\item{x}{A boolean vector}
}
\value{
An integer vector of the same length as input, providing unique trip number
}
\description{
The function
}
