% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trips.R
\name{rb_trip_jepol}
\alias{rb_trip_jepol}
\title{define_trips_jepol}
\usage{
rb_trip_jepol(
  vessel_id,
  time,
  in_harbour,
  min_dur = 0.5,
  max_dur = 72,
  split_trips = TRUE
)
}
\arguments{
\item{vessel_id}{a vector containing vessel id}

\item{time}{a vector containing timestamp}

\item{in_harbour}{a binary vector indicating if vessel in harbour (1) or not (0)}

\item{min_dur}{the minimum trip length (hours)}

\item{max_dur}{the maximum trip length (hours)}

\item{split_trips}{If the trip is longer than the maximum hours, it will try to split
the trip into two or more trips, if there is long enough intervals between pings}
}
\value{
sequential numbers identifying trips, unique within each vessel
}
\description{
Use the columns "SI_HARB" to determine when a vessel is on a trip. A trip is defined
from when it leaves the harbour till it returns
}
