% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_whacky_points}
\alias{rb_whacky_points}
\title{rb_whacky_points}
\usage{
rb_whacky_points(d, tolerance = 10)
}
\arguments{
\item{d}{A tibble containing lon and lat}

\item{tolerance}{The tolerance atomic value for marking a whacky point (TRUE), units in nautical miles}
}
\value{
A tibble with a boolean variable whacky
}
\description{
rb_whacky_points
}
\author{
Einar Hjörleifsson, \email{einar.hjorleifsson@gmail.com}
}
