% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_winsor}
\alias{rb_winsor}
\title{Calculate the upper Winsor value}
\usage{
rb_winsor(x, probs = 0.99)
}
\arguments{
\item{x}{A numeric vector}

\item{probs}{Default is 0.99}
}
\value{
The value corresponding to the percentile
}
\description{
Calculate the upper Winsor value
}
