% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{read_is_harbours}
\alias{read_is_harbours}
\title{read_is_harbours}
\usage{
read_is_harbours()
}
\value{
An sf-tibble
}
\description{
A conenient wrapper that read a dataframe of Icelandic harbours (hid) with
associated polygon geometry from an open ftp-site.
}
