% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{read_is_survey_stations}
\alias{read_is_survey_stations}
\title{read_is_survey_stations}
\usage{
read_is_survey_stations()
}
\value{
A tibble
}
\description{
A conenient wrapper that read a dataframe of Icelandic trawl vessel survey
tow data. The variables are \code{cruise_id} (cruise identification number),
\code{station_id} (unique station identification numner), \code{station_nr} 
(a non-unique station number), \code{date} (date the tow was taken),
\code{vid} (vessel identification number), \code{sclass} (sample class, 30
is the annual spring survey, 35 the annual fall survey) and time-space
informations: start (\code{lon1}, \code{lat1}, \code{t1}) and end 
(\code{lon2}, \code{lat2}, \code{t2}) of towing as reported by the vessel
captain.
}
