% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tacsat}
\alias{tacsat}
\title{Vessel Monitoring by Satellite system dataset. This dataset is from the  
vmstools with years in the dataset changed because of leap-year issues in
the original dataset.}
\format{
A data frame with 97015 observations (rows) and 9 variables (columns):
\describe{
  \item{.ridt}{Row number}
  \item{VE_REF}{Vessel id}
  \item{SI_LATI}{Latitude}
  \item{SI_LONG}{Longitude}
  \item{SI_DATE}{Date - character vector}
  \item{SI_TIME}{Time - character vector}
  \item{SI_SP}{Instantaneous speed}
  \item{SI_HE}{Instantaneous heading}
  }
}
\source{
https://doi.org/10.1093/icesjms/fsv099
}
\usage{
tacsat
}
\description{
Vessel Monitoring by Satellite system dataset. This dataset is from the  
vmstools with years in the dataset changed because of leap-year issues in
the original dataset.
}
\keyword{datasets}
