% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_returns.R
\name{tidy_bin_clst_path}
\alias{tidy_bin_clst_path}
\title{tidy_bin_clst_path}
\usage{
tidy_bin_clst_path(o)
}
\arguments{
\item{o}{Object of class ...}
}
\value{
a tibble
}
\description{
tidy_bin_clst_path
}
