% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_returns.R
\name{tidy_normalmixEM}
\alias{tidy_normalmixEM}
\title{tidy_normalmixEM}
\usage{
tidy_normalmixEM(o)
}
\arguments{
\item{o}{Object (class mixEM) that contains object of class normalmixEM}
}
\value{
a tibble
}
\description{
tidy_normalmixEM
}
