% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trawlsurvey}
\alias{trawlsurvey}
\title{Bottom trawl survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8918 rows and 7 columns.
}
\usage{
trawlsurvey
}
\description{
Icelandic spring bottom trawl survey.
}
\keyword{datasets}
