% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trawlsurveytows}
\alias{trawlsurveytows}
\title{Bottom trawl survey tows}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 581 rows and 2 columns.
}
\usage{
trawlsurveytows
}
\description{
Icelandic spring bottom trawl survey.
}
\keyword{datasets}
