% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whackies.R
\name{whack-deprecated}
\alias{whack-deprecated}
\alias{whack_clean}
\alias{whack_sda}
\alias{whack_forward}
\alias{whack_fwdbwd}
\alias{whack_sequential_fast}
\title{Deprecated names of the whacky-position filters}
\usage{
whack_clean(...)

whack_sda(...)

whack_forward(...)

whack_fwdbwd(...)

whack_sequential_fast(...)
}
\arguments{
\item{...}{Passed to the new function.}
}
\description{
`whack_clean()`, `whack_sda()`, `whack_forward()`, `whack_fwdbwd()` and
`whack_sequential_fast()` were renamed [rb_whack_clean()], [rb_whack_sda()],
[rb_whack_forward()], [rb_whack_fwdbwd()] and [rb_whack_sequential_fast()] on 2026-10-08,
in line with the rest of the package. The old names still work, with a warning.
}
\keyword{internal}
