% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_spread_cash_and_catch.R
\name{dc_spread_cash_and_catch}
\alias{dc_spread_cash_and_catch}
\title{Allocate Landings from Events to Trails}
\usage{
dc_spread_cash_and_catch(trails, events, remove_diagnostic = TRUE)
}
\arguments{
\item{trails}{A data frame containing fishing trail records. **Required columns:**
\itemize{
  \item{\code{.pid}:} {integer, unique trail identifier}
  \item{\code{VE_COU}:} {character, vessel country}
  \item{\code{VE_REF}:} {character, vessel reference}
  \item{\code{time}:} {POSIXct/datetime, UTC timestamp, derived from SI_DATE and SI_TIME}
  \item{\code{FT_REF}:} {character, fishing trip reference}
  \item{\code{ir}:} {character, ICES rectangle, derive from SI_LONG and SI_LAT}
  \item{\code{state}:} {integer, state (must be 0 or 1)}
}}
}
\value{
A data frame similar to \code{trails}, with updated \code{LE_KG}, \code{LE_EURO},
  and a column \code{.how} indicating at which step the spread took place. Additional variables 
  may be returned if arguement remove_diagnostic is set to FALSE.
}
\description{
This function allocates landing event weights (`LE_KG`) and values (`LE_EURO`)
from the `events` data frame to the corresponding fishing trails in the `trails`
data frame using a hierarchical matching and allocation strategy.
}
\details{
The function spreads event cash and catches to trails using three hierarchical steps:
\enumerate{
  \item Joins by year, trip, state, rectangle and day.
  \item Joins remaining unmatched records by year, trip, state and rectangle.
  \item Joins remaining unmatched records by year, trip, and state
}

If any trip cash or catches are remaining the above are raised by trip based raising factor.
}
