% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_rayshade_xyz_dynamic.R
\name{mb_rashade_xyz_dynamic}
\alias{mb_rashade_xyz_dynamic}
\title{Dynamic rayshading}
\usage{
mb_rashade_xyz_dynamic(
  xyz,
  r0,
  ping_cutoff = 10,
  file_prefix = "tmp/tmp",
  max_meters = 1024
)
}
\arguments{
\item{xyz}{xxx}

\item{r0}{xxx}

\item{ping_cutoff}{xxx}

\item{file_prefix}{xxx}

\item{max_meters}{Default 1024}
}
\value{
A raster
}
\description{
Dynamic rayshading
}
