% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramb-package.R
\docType{package}
\name{ramb-package}
\alias{ramb}
\alias{ramb-package}
\title{ramb: Rambling}
\description{
Some stuff on vessels tracks.
}
\author{
\strong{Maintainer}: Einar Hjörleifsson \email{einar.hjorleifsson@gmail.com}

Authors:
\itemize{
  \item Einar Hjörleifsson \email{einar.hjorleifsson@gmail.com}
}

}
\keyword{internal}
