% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramb-package.R
\docType{package}
\name{ramb-package}
\alias{ramb}
\alias{ramb-package}
\title{ramb: Rambling}
\description{
Some stuff done at WKSSFGEO 2021. On vessels tracks and things.
}
\author{
\strong{Maintainer}: Einar Hjörleifsson \email{einar.hjorleifsson@gmail.com}

}
\keyword{internal}
