% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramb-package.R
\docType{package}
\name{ramb-package}
\alias{ramb}
\alias{ramb-package}
\title{ramb: Convert "move"-object into tibbles}
\description{
The goal of {ramb} is to return various "move"-objects to tibbles for further downstream processing using the tidyverse lingo. And then some.
}
\author{
\strong{Maintainer}: Einar Hjörleifsson \email{einar.hjorleifsson@gmail.com}

}
\keyword{internal}
