% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\name{ramb-renamed}
\alias{ramb-renamed}
\alias{rb_speed}
\alias{rb_distance}
\alias{rb_st}
\alias{rb_sa}
\alias{rb_track_time}
\alias{rb_event}
\alias{rb_d2ir}
\alias{rb_midpoint}
\alias{rb_points_in_polygons}
\alias{rb_st_keep}
\alias{rb_st_drop}
\alias{rb_benthis_width}
\alias{rb_gearwidth_proxy}
\alias{rb_std_meshsize}
\alias{rb_mmsi_category}
\alias{rb_mmsi_flag}
\alias{rb_check_crosses_year}
\alias{dc_check_eflalo_crosses_year}
\alias{dc_create_timestamp}
\alias{rb_mapdeck}
\alias{rb_md_trip}
\alias{rb_summary}
\alias{rb_logbook}
\alias{rb_trail}
\alias{read_is_harbours}
\alias{read_is_survey_stations}
\alias{read_is_survey_tracks}
\alias{mb_base_raster}
\alias{mb_bb}
\alias{mb_rashade_xyz_dynamic}
\alias{mb_rasterize_xyz}
\alias{mb_rayshade_raster}
\alias{mb_rayshade_raster_rgb}
\alias{mb_rayshade_to_rgb}
\alias{mb_xyz_extent}
\alias{rb_ms2kn}
\alias{rb_kn2ms}
\alias{rb_gear_from_metier}
\alias{rb_target_from_metier}
\alias{rb_met5_from6}
\alias{rb_get_ices_gears}
\alias{rb_get_ices_target}
\alias{rb_get_ices_metier5}
\alias{rb_get_ices_metier6}
\alias{rb_get_ices_metier5_benthis_lookup}
\alias{rb_cap_iqr}
\alias{rb_cap_miller}
\alias{rb_cap_winsorize}
\title{Old function names}
\usage{
rb_speed(...)

rb_distance(...)

rb_st(...)

rb_sa(...)

rb_track_time(...)

rb_event(...)

rb_d2ir(...)

rb_midpoint(...)

rb_points_in_polygons(...)

rb_st_keep(...)

rb_st_drop(...)

rb_benthis_width(...)

rb_gearwidth_proxy(...)

rb_std_meshsize(...)

rb_mmsi_category(...)

rb_mmsi_flag(...)

rb_check_crosses_year(...)

dc_check_eflalo_crosses_year(...)

dc_create_timestamp(...)

rb_mapdeck(...)

rb_md_trip(...)

rb_summary(...)

rb_logbook(...)

rb_trail(...)

read_is_harbours(...)

read_is_survey_stations(...)

read_is_survey_tracks(...)

mb_base_raster(...)

mb_bb(...)

mb_rashade_xyz_dynamic(...)

mb_rasterize_xyz(...)

mb_rayshade_raster(...)

mb_rayshade_raster_rgb(...)

mb_rayshade_to_rgb(...)

mb_xyz_extent(...)

rb_ms2kn(...)

rb_kn2ms(...)

rb_gear_from_metier(...)

rb_target_from_metier(...)

rb_met5_from6(...)

rb_get_ices_gears(...)

rb_get_ices_target(...)

rb_get_ices_metier5(...)

rb_get_ices_metier6(...)

rb_get_ices_metier5_benthis_lookup(...)

rb_cap_iqr(...)

rb_cap_miller(...)

rb_cap_winsorize(...)
}
\arguments{
\item{...}{Passed to the new function.}
}
\description{
These names were changed on 2026-10-08 so that every function is named verb first, with a word
for its step in the fishing-activity flow (see the Naming section of \link{ramb-package}). The old
names still work, warn once per session, and will be removed in a future version.
}
\keyword{internal}
