% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{rb_cap_winsorize}
\alias{rb_cap_winsorize}
\title{Winsorize}
\usage{
rb_cap_winsorize(x, limit = stats::quantile(x, probs = c(0.99), na.rm = FALSE))
}
\arguments{
\item{x}{A numeric vector}

\item{limit}{Normally a single value (default 0.99), values higher than this will be replaced by this value.
Default is the 98th percentile of the distribution of x.}
}
\value{
A vector of same length as x
}
\description{
Replace extreme values by less extreme ones. Default is just to cap the higher
values, but if arguement p is of length 2, lower values can also be substituted.
}
