% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rb_create_timestamp}
\alias{rb_create_timestamp}
\title{Create timestamp from date and time character columns}
\usage{
rb_create_timestamp(
  date,
  time,
  tz = "GMT",
  format = "\%d/\%m/\%Y  \%H:\%M:\%S"
)
}
\arguments{
\item{date}{Name of the date variable}

\item{time}{Name of the time variable}

\item{tz}{Time zone (default GMT)}

\item{format}{time output format (default}
}
\value{
A vector of same length as date and time
}
\description{
Create timestamp from date and time character columns
}
