% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_converters.R
\name{rb_d2ir}
\alias{rb_d2ir}
\title{Get ICES rectangles from coordinates}
\usage{
rb_d2ir(lon, lat, safe = TRUE, useI = FALSE)
}
\arguments{
\item{lon}{Longitude vector}

\item{lat}{Latitude vector}

\item{safe}{Boolean (default TRUE), calculates mid-point first before calculating rectangle. Ensures consistency in binning.}

\item{useI}{Boolean (default FALSE), I is rarely used.}
}
\value{
A character vector of ICES rectangles
}
\description{
Get the ICES rectangle from any lon-lat position. If position outside
ICES region return NA's.
}
