% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_converters.R
\name{rb_grade}
\alias{rb_grade}
\title{Binning data}
\usage{
rb_grade(x, dx)
}
\arguments{
\item{x}{A numerical vector, e.g. longitude or latitude}

\item{dx}{A value, think of this a binwidth}
}
\value{
A vector with same length as x
}
\description{
Returns the midpoint of value given a binwidth (dx)
}
