% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rb_hidden_markov_step}
\alias{rb_hidden_markov_step}
\title{Hidden Markov model with step only}
\usage{
rb_hidden_markov_step(d)
}
\arguments{
\item{d}{A data frame with xx}
}
\value{
A tibble
}
\description{
Hidden Markov model with step only
}
