% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rb_hidden_markov_step_and_turn}
\alias{rb_hidden_markov_step_and_turn}
\title{Hidden Markov model with step and turn}
\usage{
rb_hidden_markov_step_and_turn(d)
}
\arguments{
\item{d}{A data frame with xx}
}
\value{
A tibble
}
\description{
Hidden Markov model with step and turn
}
