% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{rb_interval_id}
\alias{rb_interval_id}
\title{rb_interval_id}
\usage{
rb_interval_id(point, interval, vid, time, start, end, id)
}
\arguments{
\item{point}{A tibble containing variables vid (vessel id) and time}

\item{interval}{A tibble containing variables vid, id.lgs (the fishing 
activity id), start and end}

\item{vid}{The vessel id variable name}

\item{time}{The name of the time variable in the ais data}

\item{start}{The name of the start time variable in the logbooks}

\item{end}{The name of the end time variable in the logbooks}

\item{id}{The name of the tow id variable in the logbooks}
}
\value{
The point tibble with additional variable id.lgs (fishing activity
id).
}
\description{
The function aim is to match an event in one tibble with an interval in an
another tibble. Specifically designed to be used with ais/vms data (points) 
and fishing logbook records where start and end time of fishing activity is
recorded.
}
\details{
Superseeded by just using between within a dplyr-join
}
