% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_accross.R
\name{rb_sum_across}
\alias{rb_sum_across}
\title{Sum logbook species catches}
\usage{
rb_sum_across(d, start_name = NULL, remove = TRUE)
}
\arguments{
\item{d}{a eflalo dataframe}

\item{start_name}{a string, start name of variable to sum across}

\item{remove}{Boolean (default TRUE) removes catch by species}
}
\value{
Input with additional summed variable, name is 'start_name'
}
\description{
Sum logbook species catches
}
