% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whackies.R
\name{rb_whacky_distance}
\alias{rb_whacky_distance}
\title{Yet another distance filter}
\usage{
rb_whacky_distance(lon, lat, miles_max = 6)
}
\arguments{
\item{lon}{longitude in decimal degrees}

\item{lat}{latitude in decimal degrees}

\item{miles_max}{maximum threshold in miles}
}
\value{
a boolean vector, TRUE if point classified as whacky
}
\description{
To use this on multiple trips, use a grouped data frame with 
tidyverse code like 
data |>  group_by(id) |> mutate(whack = rb_whacky_distance(lon, lat))
}
\details{
The function uses traipse::track_distance, which by convention the first 
value is set to NA missing value, because the difference applies to 
each sequential pair of locations. Here, these missing values are
replaced by zero.

The function is overly liberal in that
}
