% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whackies.R
\name{rb_whacky_speed}
\alias{rb_whacky_speed}
\title{rb_whacky_speed}
\usage{
rb_whacky_speed(lon, lat, time, criteria = 20)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{time}{date-time in POSIXct}

\item{criteria}{speed in nautical miles per hour}
}
\value{
A boolean vector
}
\description{
Marks derived speed (unit knots) above a certain criterion as whacky (TRUE,
rest will be FALSE).
}
\details{
To use this on multiple vessels or trips, use a grouped data frame with 
tidyverse code 
like ...
}
