% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmendo.R
\name{rb_whacky_speed_mendo}
\alias{rb_whacky_speed_mendo}
\title{Yet another speed filter}
\usage{
rb_whacky_speed_mendo(df, speed_filter = 25)
}
\arguments{
\item{df}{A dataframe containing variables named}

\item{speed_filter}{speed in knots}
}
\value{
A filtered dataframe with a bunch of stuff
}
\description{
The input dataframe has to contain the following variables: x, y, time. device_id and seq.
The x and the y are coordates in units of meters.
}
\details{
This function is fast but it chops off equal amount of valid points as invalid
points. Function rb_whacky_speed is more conservative, albeit slower. The latter
also does not filter the data.
}
