% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whackies.R
\name{rb_whacky_speed_trip}
\alias{rb_whacky_speed_trip}
\title{Title}
\usage{
rb_whacky_speed_trip(d)
}
\arguments{
\item{d}{A dataframe contain lon, lat and time}
}
\value{
a dataframe with addition variable ".whacky"
}
\description{
Title
}
