% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whacks1}
\alias{whacks1}
\title{Fictional vessel tracks - set 1}
\format{
A data frame with 46 rows and 7 variables:
\describe{
  \item{.rid}{row number}
  \item{vid}{unique vessel identifier}
  \item{tid}{unique trip identifier}
  \item{lon}{Longitude}
  \item{whacks}{Boolean, indication if vessel on track or not}
  }
}
\usage{
whacks1
}
\description{
Some simple dataset to test speed filters
}
\keyword{datasets}
